MIGRATION

+++ src/main/java/ru/eshmakar/sweater/repos/MessageRepo.java
@@ -5,6 +5,6 @@ import ru.eshmakar.sweater.domain.Message;
 
 import java.util.List;
 
-public interface MessageRepo extends CrudRepository<Message, Integer> {
+public interface MessageRepo extends CrudRepository<Message, Long> {
     List<Message> findByTag (String tag);
 }
